lowercase version tag before checking if mariadb
This fixes a bug where MariaDB was not correctly detected as Mariadb. The reason why it was not detecting correctly is because, in my specific case the official docker image for mariadb version 11.6.2, MariaDB was identifying itself as 11.6.2-MariaDB-ubu2404 as opposed to lowercase mariadb defined in the check. This caused the deprecation warning to be shown while it should not.